Logs Book v1.0
There is the logs book project with no critical errors.

It's a basic template to start a big project with all the basic
programming logic. Just improve the design, code and functions to make
it a big project.

In code appears several times the word "post", it's because this a fork
of a miniBlog project, just assume post as log to understand better the
code.

Made by @codenoschool
Visit us at @codenoschool (GitHub, YouTube and Twitter).